+++ testing/src/main/java/io/stargate/it/http/RestApiv2Test.java
@@ -430,80 +430,6 @@ public class RestApiv2Test extends BaseIntegrationTest {
         HttpStatus.SC_NO_CONTENT);
   }
 
-  @Test
-  public void createTableWithNullOptions() throws IOException {
-    createKeyspace(keyspaceName);
-
-    TableAdd tableAdd = new TableAdd();
-    tableAdd.setName("t1");
-
-    List<ColumnDefinition> columnDefinitions = new ArrayList<>();
-
-    columnDefinitions.add(new ColumnDefinition("id", "uuid"));
-    columnDefinitions.add(new ColumnDefinition("lastName", "text"));
-    columnDefinitions.add(new ColumnDefinition("firstName", "text"));
-
-    tableAdd.setColumnDefinitions(columnDefinitions);
-
-    PrimaryKey primaryKey = new PrimaryKey();
-    primaryKey.setPartitionKey(Collections.singletonList("id"));
-    tableAdd.setPrimaryKey(primaryKey);
-    tableAdd.setTableOptions(null);
-
-    String body =
-        RestUtils.post(
-            authToken,
-            String.format("%s:8082/v2/schemas/keyspaces/%s/tables", host, keyspaceName),
-            objectMapper.writeValueAsString(tableAdd),
-            HttpStatus.SC_CREATED);
-
-    NameResponse response = objectMapper.readValue(body, NameResponse.class);
-    assertThat(response.name).isEqualTo(tableAdd.getName());
-  }
-
-  @Test
-  public void createTableMissingClustering() throws IOException {
-    createKeyspace(keyspaceName);
-    TableAdd tableAdd = new TableAdd();
-    tableAdd.setName(tableName);
-
-    List<ColumnDefinition> columnDefinitions = new ArrayList<>();
-
-    columnDefinitions.add(new ColumnDefinition("pk1", "int"));
-    columnDefinitions.add(new ColumnDefinition("ck1", "int"));
-
-    tableAdd.setColumnDefinitions(columnDefinitions);
-
-    PrimaryKey primaryKey = new PrimaryKey();
-    primaryKey.setPartitionKey(Collections.singletonList("pk1"));
-    primaryKey.setClusteringKey(Collections.singletonList("ck1"));
-    tableAdd.setPrimaryKey(primaryKey);
-
-    tableAdd.setTableOptions(new TableOptions(0, null));
-
-    String body =
-        RestUtils.post(
-            authToken,
-            String.format("%s:8082/v2/schemas/keyspaces/%s/tables", host, keyspaceName),
-            objectMapper.writeValueAsString(tableAdd),
-            HttpStatus.SC_CREATED);
-
-    NameResponse response = objectMapper.readValue(body, NameResponse.class);
-    assertThat(response.name).isEqualTo(tableAdd.getName());
-
-    body =
-        RestUtils.get(
-            authToken,
-            String.format(
-                "%s:8082/v2/schemas/keyspaces/%s/tables/%s?raw=true",
-                host, keyspaceName, tableName),
-            HttpStatus.SC_OK);
-
-    TableResponse table = objectMapper.readValue(body, TableResponse.class);
-    assertThat(table.getTableOptions().getClusteringExpression().get(0).getOrder())
-        .isEqualTo("ASC");
-  }
-
   /*
   /************************************************************************
   /* Test methods for Row Access CRUD operations
@@ -2881,67 +2807,6 @@ public class RestApiv2Test extends BaseIntegrationTest {
     assertThat(fields.size()).isEqualTo(1);
     assertThat(fields.get(0).get("name")).isEqualTo("firstname");
     assertThat(fields.get(0).get("typeDefinition")).isEqualTo("text");
-  }
-
-  @Test
-  public void testMixedCaseTable() throws IOException {
-    createKeyspace(keyspaceName);
-
-    String tableName = "MixedCaseTable";
-    TableAdd tableAdd = new TableAdd();
-    tableAdd.setName(tableName);
-
-    List<ColumnDefinition> columnDefinitions = new ArrayList<>();
-
-    columnDefinitions.add(new ColumnDefinition("ID", "uuid"));
-    columnDefinitions.add(new ColumnDefinition("Lastname", "text"));
-    columnDefinitions.add(new ColumnDefinition("Firstname", "text"));
-    tableAdd.setColumnDefinitions(columnDefinitions);
-
-    PrimaryKey primaryKey = new PrimaryKey();
-    primaryKey.setPartitionKey(Collections.singletonList("ID"));
-    tableAdd.setPrimaryKey(primaryKey);
-
-    // ensure table name is preserved
-    String body =
-        RestUtils.post(
-            authToken,
-            String.format("%s:8082/v2/schemas/keyspaces/%s/tables", host, keyspaceName),
-            objectMapper.writeValueAsString(tableAdd),
-            HttpStatus.SC_CREATED);
-
-    TableResponse tableResponse =
-        objectMapper.readValue(body, new TypeReference<TableResponse>() {});
-    assertThat(tableResponse.getName()).isEqualTo(tableName);
-
-    // insert a row
-    String rowIdentifier = UUID.randomUUID().toString();
-    Map<String, String> row = new HashMap<>();
-    row.put("ID", rowIdentifier);
-    row.put("Firstname", "John");
-    row.put("Lastname", "Doe");
-
-    RestUtils.post(
-        authToken,
-        String.format("%s:8082/v2/keyspaces/%s/%s", host, keyspaceName, tableName),
-        objectMapper.writeValueAsString(row),
-        HttpStatus.SC_CREATED);
-
-    // retrieve the row by ID and ensure column names are as expected
-    String whereClause = String.format("{\"ID\":{\"$eq\":\"%s\"}}", rowIdentifier);
-    body =
-        RestUtils.get(
-            authToken,
-            String.format(
-                "%s:8082/v2/keyspaces/%s/%s?where=%s", host, keyspaceName, tableName, whereClause),
-            HttpStatus.SC_OK);
-
-    ListOfMapsGetResponseWrapper getResponseWrapper =
-        LIST_OF_MAPS_GETRESPONSE_READER.readValue(body);
-    List<Map<String, Object>> data = getResponseWrapper.getData();
-    assertThat(data.get(0).get("ID")).isEqualTo(rowIdentifier);
-    assertThat(data.get(0).get("Firstname")).isEqualTo("John");
-    assertThat(data.get(0).get("Lastname")).isEqualTo("Doe");
   }
 
   /*
